Revert "refactor: 러닝 종료 UI 컴포넌트 분리 및 저장 로직 구현"
