+++ src/components/MapContainer.module.css
@@ -12,21 +12,6 @@
   border: 1px solid #ccc;
 }
 
-/* 타이머 위치 및 스타일 */
-.timer {
-  position: absolute;
-  top: 20px;
-  left: 20px;
-  background-color: rgba(0, 0, 0, 0.6);
-  color: white;
-  padding: 8px 16px;
-  border-radius: 8px;
-  font-size: 16px;
-  font-weight: bold;
-  z-index: 10;
-  user-select: none;
-}
-
 /* 공통 버튼 스타일 */
 .runButton,
 .stopButton {
@@ -63,24 +48,3 @@
 .stopButton:hover {
   background-color: #c82333;
 }
-
-.summaryBox {
-  background-color: white;
-  padding: 1rem;
-  border-radius: 1rem;
-  box-shadow: 0 0 10px rgba(0,0,0,0.1);
-  position: absolute;
-  bottom: 80px;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 10;
-  text-align: center;
-}
-
-.buttonGroup {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  margin-top: 1rem;
-}
-
