--- src/components/buttons/StartButton.jsx
@@ -1,9 +1,0 @@
-import styles from "../MapContainer.module.css";
-
-const StartButton = ({ onClick }) => (
-  <button onClick={onClick} className={styles.runButton}>
-    🏃 러닝 시작
-  </button>
-);
-
-export default StartButton;
