--- src/components/buttons/StopButton.jsx
@@ -1,9 +1,0 @@
-import styles from "../MapContainer.module.css";
-
-const StopButton = ({ onClick }) => (
-  <button onClick={onClick} className={styles.stopButton}>
-    🛑 러닝 종료
-  </button>
-);
-
-export default StopButton;
