--- src/pages/myRecords.module.css
@@ -1,73 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 2rem;
-  font-family: "Arial", sans-serif;
-}
-
-.header {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 2rem;
-}
-
-.profileSection {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-bottom: 2rem;
-}
-
-.profileSection img {
-  width: 80px;
-  height: 80px;
-  border-radius: 50%;
-  background-color: #eee;
-}
-
-.profileSection button {
-  margin-top: 0.5rem;
-  padding: 0.5rem 1rem;
-  font-size: 14px;
-}
-
-.recordList {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  justify-content: center;
-  width: 100%;
-}
-
-.recordItem {
-  border: 1px solid #ddd;
-  border-radius: 12px;
-  width: 160px;
-  padding: 0.5rem;
-  background-color: #fff;
-  text-align: center;
-  box-shadow: 0 2px 6px rgba(0,0,0,0.05);
-  cursor: pointer;
-}
-
-.recordItem img {
-  width: 100%;
-  height: 100px;
-  object-fit: cover;
-  margin-bottom: 0.5rem;
-}
-
-.recordItem .date {
-  font-size: 12px;
-  color: #666;
-  margin-bottom: 0.5rem;
-}
-
-.bottomNav {
-  margin-top: 2rem;
-  display: flex;
-  justify-content: center;
-  gap: 2rem;
-}
