--- src/utils/timeUtils.js
@@ -1,7 +1,0 @@
-// src/utils/timeUtils.js
-export const formatElapsedTime = (seconds) => {
-  const h = Math.floor(seconds / 3600);
-  const m = Math.floor((seconds % 3600) / 60);
-  const s = seconds % 60;
-  return [h, m, s].map((v) => String(v).padStart(2, "0")).join(":");
-};
